% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdgc.R
\name{get_mdgc_log_ml}
\alias{get_mdgc_log_ml}
\alias{get_mdgc_log_ml.mdgc}
\alias{get_mdgc_log_ml.data.frame}
\alias{get_mdgc_log_ml.default}
\title{Get Pointer to C++ Object to Approximate the Log Marginal Likelihood}
\usage{
get_mdgc_log_ml(object, ...)

\method{get_mdgc_log_ml}{mdgc}(object, ...)

\method{get_mdgc_log_ml}{data.frame}(object, ...)

\method{get_mdgc_log_ml}{default}(object, lower, upper, code, ...)
}
\arguments{
\item{object}{mdgc object from \code{\link{get_mdgc}} or a
\code{\link{data.frame}} to pass to \code{\link{get_mdgc}}. Ignored by
the default method.}

\item{...}{used to pass arguments to S3 methods.}

\item{lower}{#variables x #observation matrix with lower bounds
for each variable on the normal scale.}

\item{upper}{#variables x #observation matrix with upper bounds
for each variable on the normal scale.}

\item{code}{#variables x #observation matrix integer code for the
each variable on the normal scale. Zero implies an observed value (the
value in \code{upper}), one implies a missing value, and two implies an
interval.}
}
\description{
Creates a C++ object which is needed to approximate the log marginal
likelihood. The object cannot be saved.
}
\seealso{
\code{\link{mdgc_fit}}, \code{\link{mdgc_log_ml}}
}
