% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdgc-package.R
\docType{package}
\name{mdgc-package}
\alias{mdgc-package}
\alias{_PACKAGE}
\title{mdgc: Missing Data imputation using Gaussian Copulas}
\description{
The mdgc package is used to estimate Gaussian Copula models for mixed data
types (continuous, binary, ordinal, and multinomial) that can be used for
imputations. The main function is the \code{\link{mdgc}} function. The rest
of the functions in the package give the user access to lower level
functions.

Examples are provided at \url{https://github.com/boennecd/mdgc}. The package
is still in a development stage and the API may change.
}
\references{
Christoffersen, B., Clements, M., Humphreys, K., & Kjellström, H. (2021).
\emph{Asymptotically Exact and Fast Gaussian Copula Models for Imputation of Mixed Data Types}.
\url{https://arxiv.org/abs/2102.02642}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/boennecd/mdgc}
  \item Report bugs at \url{https://github.com/boennecd/mdgc/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Christoffersen \email{boennecd@gmail.com} (\href{https://orcid.org/0000-0002-7182-1346}{ORCID})

Other contributors:
\itemize{
  \item Alan Genz [copyright holder]
  \item Frank Bretz [copyright holder]
  \item Torsten Hothorn [copyright holder]
  \item R-core \email{R-core@R-project.org} [copyright holder]
  \item Ross Ihaka [copyright holder]
}

}
