% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdgc.R
\name{mdgc_fit}
\alias{mdgc_fit}
\title{Estimate the Correlation Matrix}
\usage{
mdgc_fit(
  ptr,
  vcov,
  lr = 0.001,
  rel_eps = 0.01,
  maxit = 10L,
  batch_size = NULL,
  method = c("svrg", "adam"),
  seed = 1L,
  epsilon = 1e-08,
  beta_1 = 0.9,
  beta_2 = 0.999,
  n_threads = 1L,
  do_reorder = TRUE,
  abs_eps = -1,
  maxpts = 100000L,
  minvls = 100L
)
}
\arguments{
\item{ptr}{returned object from \code{\link{get_mdgc_log_ml}}.}

\item{vcov}{starting value.}

\item{lr}{learning rate.}

\item{rel_eps}{relative error for each term.}

\item{maxit}{maximum number of iteration.}

\item{batch_size}{number of observations in each batch.}

\item{method}{estimation method to use.}

\item{seed}{fixed seed to use. Use \code{NULL} if the seed should not be
fixed.}

\item{epsilon, beta_1, beta_2}{ADAM parameters.}

\item{n_threads}{number of threads to use.}

\item{do_reorder}{logical for whether to use heuristic variable reordering.}

\item{abs_eps}{absolute convergence threshold for each term.}

\item{maxpts}{maximum number of samples to draw.}

\item{minvls}{minimum number of samples.}
}
\description{
Estimate the Correlation Matrix
}
