% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdgc.R
\name{mdgc_start_value}
\alias{mdgc_start_value}
\alias{mdgc_start_value.default}
\title{Get Starting Value for the Correlation Matrix Using a Heuristic}
\usage{
mdgc_start_value(object, ...)

mdgc_start_value(object, ...)

mdgc_start_value.default(
  object,
  lower,
  upper,
  code,
  categorical,
  n_threads = 1L,
  ...
)
}
\arguments{
\item{object}{mdgc object from \code{\link{get_mdgc}} or a
\code{\link{data.frame}} to pass to \code{\link{get_mdgc}}. Ignored by
the default method.}

\item{...}{used to pass arguments to S3 methods.}

\item{lower}{#variables x #observation matrix with lower bounds
for each variable on the normal scale.}

\item{upper}{#variables x #observation matrix with upper bounds
for each variable on the normal scale.}

\item{code}{#variables x #observation matrix integer code for the
each variable on the normal scale. Zero implies an observed value (the
value in \code{upper}), one implies a missing value, and two implies an
interval.}

\item{categorical}{\code{\link{list}} with 3xn \code{\link{matrix}} with
categorical outcomes. The first index is the outcome as an integer code,
the second index is the number of categories, and the third index is the
index of each categorial variable.}

\item{n_threads}{number of threads to use.}
}
\description{
Uses a heuristic to get starting values for the correlation matrix. These
can be passed e.g. to \code{\link{mdgc_fit}}.
}
