% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdgc.R
\name{mgdc_fit}
\alias{mgdc_fit}
\title{Estimates the Correlation Matrix}
\usage{
mgdc_fit(
  ptr,
  vcov,
  lr = 0.001,
  maxit = 10L,
  batch_size = NULL,
  method = c("svrg", "adam"),
  seed = 1L,
  epsilon = 1e-08,
  beta_1 = 0.9,
  beta_2 = 0.999,
  n_threads = 4L
)
}
\arguments{
\item{ptr}{returned object from \code{\link{get_mgdc_log_ml}}.}

\item{vcov}{starting value.}

\item{lr}{learning rate.}

\item{maxit}{maximum number of iteration.}

\item{batch_size}{number of observations in each batch.}

\item{method}{estimation method to use.}

\item{seed}{fixed seed to use.}

\item{epsilon, beta_1, beta_2}{ADAM parameters.}

\item{n_threads}{number of threads to use.}
}
\description{
Estimates the Correlation Matrix
}
