% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdgc.R
\name{mgdc_log_ml}
\alias{mgdc_log_ml}
\title{Evaluate the Log Marginal Likelihood and Its Derivatives}
\usage{
mgdc_log_ml(
  ptr,
  vcov,
  releps = 0.01,
  n_threads = 4L,
  comp_derivs = FALSE,
  indices = NULL,
  do_reorder = TRUE,
  maxpts = 100000L,
  abseps = -1
)
}
\arguments{
\item{ptr}{object returned by \code{\link{get_mgdc_log_ml}}.}

\item{vcov}{correlation matrix.}

\item{releps}{relative error for each term.}

\item{n_threads}{number of threads.}

\item{comp_derivs}{logical for whether to approximate the gradient.}

\item{indices}{integer vector with which terms to include. Must be zero
indexed. \code{NULL} yields all observations.}

\item{do_reorder}{logical for whether to use heuristic variable reordering.}

\item{maxpts}{maximum number of samples to draw.}

\item{abseps}{absolute convergence threshold for each term.}
}
\description{
Evaluate the Log Marginal Likelihood and Its Derivatives
}
