% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdgc.R
\name{get_mdgc}
\alias{get_mdgc}
\title{Get mdgc Object}
\usage{
get_mdgc(dat)
}
\arguments{
\item{dat}{\code{\link{data.frame}} with continuous, ordinal, and binary
data.}
}
\description{
Creates a mdgc object which is needed for estimation of the
covariance matrix and to perform imputation.
}
\seealso{
\code{\link{get_mdgc_log_ml}}, \code{\link{mdgc_impute}}
}
