% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdgc.R
\name{get_mdgc_log_ml}
\alias{get_mdgc_log_ml}
\alias{get_mdgc_log_ml.mdgc}
\alias{get_mdgc_log_ml.default}
\title{Get Pointer to C++ Object to Approximate the Log Marginal Likelihood}
\usage{
get_mdgc_log_ml(object, ...)

\method{get_mdgc_log_ml}{mdgc}(object, ...)

\method{get_mdgc_log_ml}{default}(object, lower, upper, code, ...)
}
\arguments{
\item{object}{mdgc object from \code{\link{get_mdgc}}. Ignored by the
default method.}

\item{...}{used to pass arguments to S3 methods.}

\item{lower}{#variables x #observation matrix with lower bounds
for each variable on the normal scale.}

\item{upper}{#variables x #observation matrix with upper bounds
for each variable on the normal scale.}

\item{code}{#variables x #observation matrix integer code for the
each variable on the normal scale. Zero implies an observed value,
one implies a missing value, and two implies an interval.}
}
\description{
Get Pointer to C++ Object to Approximate the Log Marginal Likelihood
}
