% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdgc.R
\name{mdgc_impute}
\alias{mdgc_impute}
\title{Impute Missing Values Given a Covariance Matrix}
\usage{
mdgc_impute(
  object,
  vcov,
  mea,
  rel_eps = 0.001,
  maxit = 10000L,
  abs_eps = -1,
  n_threads = 1L,
  do_reorder = TRUE,
  minvls = 1000L,
  use_aprx = FALSE
)
}
\arguments{
\item{object}{returned object from \code{\link{get_mdgc}}.}

\item{vcov}{covariance matrix to condition on in the imputation.}

\item{mea}{vector with non-zero mean entries.}

\item{rel_eps}{relative error for each term.}

\item{maxit}{maximum number of iteration.}

\item{abs_eps}{absolute convergence threshold for each term.}

\item{n_threads}{number of threads to use.}

\item{do_reorder}{logical for whether to use a heuristic variable
reordering. \code{TRUE} is likely the best option.}

\item{minvls}{minimum number of samples.}

\item{use_aprx}{logical for whether to use an approximation of
\code{\link{pnorm}} and \code{\link{qnorm}}. This may yield a
noticeable reduction in the computation time.}
}
\value{
A list with imputed values for the continuous variables and a vector with
probabilities for each level for the ordinal variables.
}
\description{
Imputes missing values given a covariance matrix using a similar
quasi-random numbers method as \code{\link{mdgc_fit}}.
}
