% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdgc.R
\name{mdgc_log_ml}
\alias{mdgc_log_ml}
\title{Evaluate the Log Marginal Likelihood and Its Derivatives}
\usage{
mdgc_log_ml(
  ptr,
  vcov,
  rel_eps = 0.01,
  n_threads = 1L,
  comp_derivs = FALSE,
  indices = NULL,
  do_reorder = TRUE,
  maxpts = 100000L,
  abs_eps = -1,
  minvls = 100L,
  use_aprx = FALSE
)
}
\arguments{
\item{ptr}{object returned by \code{\link{get_mdgc_log_ml}}.}

\item{vcov}{correlation matrix.}

\item{rel_eps}{relative error for each term.}

\item{n_threads}{number of threads to use.}

\item{comp_derivs}{logical for whether to approximate the gradient.}

\item{indices}{integer vector with which terms (observations) to include.
Must be zero indexed. \code{NULL} yields all observations.}

\item{do_reorder}{logical for whether to use heuristic variable
reordering. \code{TRUE} is likely the best option.}

\item{maxpts}{maximum number of samples to draw.}

\item{abs_eps}{absolute convergence threshold for each term.}

\item{minvls}{minimum number of samples.}

\item{use_aprx}{logical for whether to use an approximation of
\code{\link{pnorm}}.}
}
\description{
Approximates the log marginal likelihood and the derivatives using
quasi-random numbers.
}
\seealso{
\code{\link{mdgc_fit}}
}
